feat(ai): add help command for AI interactions
