+++ pkg/discord/view/ai/interface.go
@@ -6,4 +6,5 @@ import (
 
 type AIViewer interface {
 	SendResponse(message *model.DiscordMessage, response *model.AIResponse) error
+	Help(message *model.DiscordMessage) error
 }
